use 3 digit version starting with Moodle 400
