+++ Moosh/Command/Moodle401/Question/QuestionDelete.php
@@ -6,7 +6,7 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace Moosh\Command\Moodle41\Question;
+namespace Moosh\Command\Moodle401\Question;
 
 use Moosh\MooshCommand;
